+++ src/com/github/gumtreediff/matchers/heuristic/gt/ParentsMappingComparator.java
@@ -42,7 +42,7 @@ public final class ParentsMappingComparator extends AbstractMappingComparator {
     protected double parentsJaccardSimilarity(ITree src, ITree dst) {
         List<ITree> srcParents = src.getParents();
         List<ITree> dstParents = dst.getParents();
-        double numerator = (double) StringAlgorithms.lcss(srcParents, dstParents).size();
+        double numerator = StringAlgorithms.lcss(srcParents, dstParents).size();
         double denominator = (double) srcParents.size() + (double) dstParents.size() - numerator;
         return numerator / denominator;
     }
